Add files via upload
A bash script for analyzing aws auth.logs to search for brute force attempts. Code pulls up IP addresses trying to login with an invalid username as well as a list of invalid usernames, IP addresses that attempt to login to root, and IP addresses that attempt to login to Ubuntu.